% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{AUC: Area Under the Curve of the Receiver Operating Curve}
\usage{
auc(obs, pred)
}
\arguments{
\item{obs}{a vector of observed values which must be 0 for absences and 1 for occurrences.}

\item{pred}{a vector of the same length as \code{obs} representing the predicted values. Values must be between 0 & 1.}
}
\value{
Returns a single value representing the AUC value.
}
\description{
Estimates the AUC of the ROC using a Mann-Whitney U statistic. Excludes missing data, if any.
}
\examples{
# create some data
obs = c(sample(c(0,1),20,replace=TRUE),NA)
pred = runif(length(obs),0,1)

# calculate AUC from the random data
auc(obs,pred)

# calculate an example 'perfect' AUC
obs = obs[order(obs)]
pred = pred[order(pred)]
auc(obs,pred)
}
\author{
Jeremy VanDerWal
}
\seealso{
Package \code{SDMTools}
}

