% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dropbox_path.R
\name{dropbox_path}
\alias{dropbox_path}
\title{Get path to local Dropbox folder.}
\usage{
dropbox_path()
}
\value{
A character vector with the path to the Dropbox folder.
}
\description{
This function returns the full path to the Dropbox folder in the local system. Currently it works for Windows only.
}
\author{
Petr Simecek
}
\seealso{
\link{http://applyr.blogspot.com.es/2012/08/get-path-to-your-dropbox-folder.html}.
}

