% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gfitLogReg.R
\name{gfitLogReg}
\alias{gfitLogReg}
\title{Goodness of fit of a Logistic Regression}
\usage{
gfitLogReg(obs, pred, groups, print.table = FALSE, main = "Goodness of fit",
  xlab = "Observations", ylab = "Predictions", xlim = c(0, 1), pch = 16,
  ps = 2, cex.lab = 1.2, cex.axis = 1.1, las = 1, ...)
}
\arguments{
\item{obs}{numeric vector with observations (either 0 or 1)}

\item{pred}{numeric vector (same length as obs) with fitted model probabilities}

\item{groups}{Number of groups (bins) to divide the data into. If missing, the default is 10.}

\item{print.table}{Logical (default is FALSE).}

\item{...}{other arguments (e.g. xlim, xlab, main, ...) to be passed to \code{plot}. Some default values have been chosen.}
}
\value{
A goodness of fit plot. Optionally (if \code{print.table=T}), also a table. See \code{plotCalibration} from package \code{PredictABEL}.
}
\description{
Make a goodness of fit plot (observed vs predicted values) in a logistic regression context. Optionally, a table with
numerical results (including a chi-squared test) can be produced.
}
\examples{
# Generate some data
x <- seq(1:1000)
yhat <- plogis(0.2 + 0.003*x)   # inverse logit
y <- rbinom(1000, 1, yhat)

# Fit model and get predicted probabilities
model <- glm(y~x, family="binomial")
ypred <- predict(model, type="response")

# Check goodness of fit
gfitLogReg(y, ypred)
gfitLogReg(y, ypred, print.table=TRUE)
gfitLogReg(y, ypred, xlim=c(0.5,1))
gfitLogReg(y, ypred, groups=5)
}
\author{
Paco, based on function \code{plotCalibration} from package \code{PredictABEL}.
}
\seealso{
\code{plotCalibration} from package \code{PredictABEL}
}

