% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_columns.R
\name{move_columns}
\alias{move_columns}
\title{Move the columns of a dataframe relative to each other.}
\usage{
move_columns(x, who, before = NULL, after = NULL)
}
\arguments{
\item{x}{a dataframe.}

\item{who}{A character vector of column names to move, or a logical vector of length names(x), or a vector of indices}

\item{before}{the column before which to put \code{who}: may be character, numeric or NA. If NA, the named columns are moved to the front (before the first column).}

\item{after}{the column after which to put \code{who}: may be character, integer, or NA. If NA, the named columns are moved to the end (after the last column).}
}
\value{
A data frame with shuffled columns.
}
\description{
Move the columns of a dataframe relative to each other.
}
\note{
This function has been modified from archived package \code{metrumrg} where it was named \code{shuffle}.
}
\examples{
\dontrun{
data(iris)
iris.new <- move_columns(iris, "Species", after = NA) # move species to last column
iris.new <- move_columns(iris, "Species", after = 2)
iris.new <- move_columns(iris, "Species", after = "Sepal.Length")
iris.new <- move_columns(iris, "Species", before = NA) # move species to first column
iris.new <- move_columns(iris, "Species", before = 3)
iris.new <- move_columns(iris, "Species", before = "Petal.Length")
}
}
\seealso{
\link{https://github.com/cran/metrumrg/blob/master/R/shuffle.R}.
}
