% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r2.gelman.R
\name{r2.gelman}
\alias{r2.gelman}
\title{Calculate R squared (R2) for a multilevel/hierarchical model following Gelman & Pardoe (2006)}
\usage{
r2.gelman(residuals, data.level, responsevar, param.sims)
}
\arguments{
\item{residuals}{A numeric matrix containing the residual values obtained through MCMC.}

\item{data.level}{Logical. If TRUE, calculates the R2 at the data level (level 1) and the denominator is the variance of the data
instead of variance of the parameters/linear predictors (see function definition for more details).}

\item{responsevar}{A numeric vector. The response variable used to fit the model. Only required for data-level calculations.}

\item{param.sims}{A numeric matrix containing the posterior samples for the random effects parameters.}
}
\value{
A numeric value: the r-squared
}
\description{
This function calculates the R-squared of a multilevel/hierarchical model following Gelman & Pardoe (2006) (see \code{References}).
}
\author{
Paco, based on code by Gelman & Hill (2007) (see \code{references}).
}
\references{
Gelman & Pardoe (2006) Technometrics. Gelman & Hill (2007) book
}
\seealso{
\code{\link{r2.multilevel}} in this same package.

\code{\link{r.squaredGLMM}} in \code{MuMIn} package and \link{http://jslefche.wordpress.com/2013/03/13/r2-for-linear-mixed-effects-models/}
}

