% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/roxytemplate.R
\name{roxytemplate}
\alias{roxytemplate}
\title{Populate the boilerplate roxygen template at the top of the function.}
\usage{
roxytemplate(funfile, params_start = 1, params_end = NULL)
}
\arguments{
\item{funfile}{path to the .R file containing the function}

\item{params_start}{The (first) line that contains the parameters for your function (default 1)}

\item{params_end}{(optional) If your parameter definitions breaks across multiple lines,
specify the ending line (default \code{NULL}).}
}
\value{
nothing, but adds the roxygen template to the top of the file
}
\description{
Populate the boilerplate roxygen template at the top of the function.
}
\author{
A. Teucher. Inspired by Karthik Ram's RTools Sublime Text 2 plugin:
https://github.com/karthik/Rtools
}

