\name{savepdfSmallMarg}
\alias{savepdfSmallMarg}
\title{Save plot as a pdf file with reduced margins (suitable for publication)}
\usage{
  savepdfSmallMarg(file, width = 16, height = 10)
}
\arguments{
  \item{file}{filename}

  \item{width}{width in cm}

  \item{height}{height in cm}
}
\value{
  a pdf file in the working directory
}
\description{
  Save plot as a pdf file with reduced margins (suitable
  for publication)
}
\author{
  Rob Hyndman
}

