% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/words_chars.R
\name{words}
\alias{chars}
\alias{words}
\title{Word and character counts}
\usage{
words(text)

chars(text, spaces = FALSE)
}
\arguments{
\item{text}{Text to count words or characters in. If missing, will read from the clipboard.}

\item{spaces}{Logical. Include spaces? Default is FALSE.}
}
\description{
Counts number of words and characters (with or without spaces) in a piece of text.
}
\examples{
\dontrun{
# Copy text to clipboard and then run
words()
chars()
chars(spaces = TRUE)

# Or provide text:
words("This is just a trial")
}
}
\author{
F. Rodriguez-Sanchez.
}

