% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacotools.R
\docType{package}
\name{pacotools}
\alias{pacotools}
\alias{pacotools-package}
\title{A collection of miscellaneous tools I often need and forget about}
\description{
More generally, an attempt to stop reinventing the wheel every monday
}
\author{
F. Rodriguez-Sanchez, with some functions collected from different authors
}
\seealso{
\code{arm} package
}
\keyword{package}

